+++ src/messenger/components/ConversationListItem/ConversationListItem.css
@@ -9,6 +9,11 @@
   cursor: pointer;
 }
 
+.conversation-list-item:active {
+  background: grey;
+  cursor: pointer;
+}
+
 .conversation-photo {
   width: 50px;
   height: 50px;
